Use the existing endpoint when binding to existing socket
UdtSharp tried to get the local endpoint from an existing socket from its
(i.e. UdtSharps) internal structure. This is wrong since the internal structure
for the socket has not been setup at this point which caused an exception.
This method was unused before which is probably the cause that the error
has not been fixed yet.
We use the sockets `.LocalEndPoint` to get the endpoint.