Add a method to bind to an existing socket
We need to be able to reuse an exisiting socket because we contacted a STUN
server which gave us our external endpoint. If we were to create a new socket
the external endpoint would change.
There was no way to bind a `UdtSocket` to an existing UDP socket.
A method `Bind` has been added which takes a .NET socket as argument.